Import updates

+++ Http/Controllers/Stream/ManagementController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\Schema;
 use Modules\AEGIS\Imports\DocumentSignatureImport;
 use Modules\AEGIS\Imports\DocumentsImport;
 use Modules\AEGIS\Imports\ProjectsImport;
+use Modules\AEGIS\Imports\RetryDocumentSignatureImport;
 use Modules\AEGIS\Imports\SignatureImport;
 use Modules\AEGIS\Imports\StoreImport;
 use Modules\AEGIS\Imports\UsersImport;
@@ -26,6 +27,9 @@ class ManagementController extends Controller
 
     public function import(SSEStream $stream, Request $request)
     {
+        // $mode = 'collect';
+        $mode = 'store';
+        // $mode = 'both';
         ini_set('max_execution_time', 0);
         $stream->send([
             'percentage' => 0,
@@ -34,41 +38,48 @@ class ManagementController extends Controller
         // Prepare blank .JSON
         $this->import_file_path = \Module::getModulePath('AEGIS').'/Resources/files/import/';
 
-        // Schema::disableForeignKeyConstraints();
-        // foreach ([
-        //     '\Modules\AEGIS\Models\CompanyType',
-        //     '\Modules\AEGIS\Models\DocumentApprovalItemDetails',
-        //     '\Modules\AEGIS\Models\FeedbackListType',
-        //     '\Modules\AEGIS\Models\JobTitle',
-        //     '\Modules\AEGIS\Models\Project',
-        //     '\Modules\AEGIS\Models\ProjectVariant',
-        //     '\Modules\AEGIS\Models\Scope',
-        //     '\Modules\AEGIS\Models\Type',
-        //     '\Modules\AEGIS\Models\VariantDocument',
+        if (in_array($mode, ['both', 'store'])) {
+            Schema::disableForeignKeyConstraints();
+            foreach ([
+                '\App\Models\Activity',
 
-        //     '\Modules\Documents\Models\ApprovalItemGroup',
-        //     '\Modules\Documents\Models\ApprovalProcess',
-        //     '\Modules\Documents\Models\ApprovalProcessItem',
-        //     '\Modules\Documents\Models\ApprovalProcessStage',
-        //     '\Modules\Documents\Models\Category',
-        //     '\Modules\Documents\Models\CategoryApprovalProcess',
-        //     '\Modules\Documents\Models\CategoryApprovalProcessItem',
-        //     '\Modules\Documents\Models\Comment',
-        //     '\Modules\Documents\Models\Document',
-        //     '\Modules\Documents\Models\DocumentApprovalProcessItem',
-        //     '\Modules\Documents\Models\Group',
-        //     '\Modules\Documents\Models\UserGroup',
-        // ] as $class) {
-        //     $class::truncate();
-        // }
-        // Schema::enableForeignKeyConstraints();
-        \Storage::delete([
-            'modules/aegis/import/projects_and_documents.json',
-            'modules/aegis/import/projects_and_document_signatures.json',
-            'modules/aegis/import/project_data.json',
-        ]);
-        \Storage::put('modules/aegis/import/errors.json', json_encode([]));
-        \Storage::put('modules/aegis/import/projects.json', json_encode([]));
+                '\Modules\AEGIS\Models\CompanyType',
+                '\Modules\AEGIS\Models\DocumentApprovalItemDetails',
+                '\Modules\AEGIS\Models\FeedbackListType',
+                '\Modules\AEGIS\Models\JobTitle',
+                '\Modules\AEGIS\Models\Project',
+                '\Modules\AEGIS\Models\ProjectVariant',
+                '\Modules\AEGIS\Models\Scope',
+                '\Modules\AEGIS\Models\Type',
+                '\Modules\AEGIS\Models\VariantDocument',
+
+                '\Modules\Documents\Models\ApprovalItemGroup',
+                '\Modules\Documents\Models\ApprovalProcess',
+                '\Modules\Documents\Models\ApprovalProcessItem',
+                '\Modules\Documents\Models\ApprovalProcessStage',
+                '\Modules\Documents\Models\Category',
+                '\Modules\Documents\Models\CategoryApprovalProcess',
+                '\Modules\Documents\Models\CategoryApprovalProcessItem',
+                '\Modules\Documents\Models\Comment',
+                '\Modules\Documents\Models\Document',
+                '\Modules\Documents\Models\DocumentApprovalProcessItem',
+                '\Modules\Documents\Models\Group',
+                '\Modules\Documents\Models\UserGroup',
+            ] as $class) {
+                $class::truncate();
+            }
+            Schema::enableForeignKeyConstraints();
+        }
+        if (in_array($mode, ['both', 'collect'])) {
+            \Storage::delete([
+                'modules/aegis/import/retry_document_signatures.json',
+                'modules/aegis/import/projects_and_documents.json',
+                'modules/aegis/import/projects_and_document_signatures.json',
+                'modules/aegis/import/project_data.json',
+            ]);
+            \Storage::put('modules/aegis/import/errors.json', json_encode([]));
+            \Storage::put('modules/aegis/import/projects.json', json_encode([]));
+        }
         $steps = [
             'users' => [
                 'users.xlsx',
@@ -76,23 +87,30 @@ class ManagementController extends Controller
             'document_types' => [
                 'processes.php',
             ],
-            'projects' => [
-                'acs/ALL_PROJECTS.xlsx',
-                'aes/ALL_PROJECTS.xlsx',
-            ],
-            'documents' => [
-                'acs/DOCUMENTS.xlsx',
-                'aes/DOCUMENTS.xlsx',
-            ],
-            'document_signatures' => [
-                'acs/DOCUMENTS_Signature.xlsx',
-                'aes/DOCUMENTS_Signature.xlsx',
-            ],
-            'signatures' => [
-                'acs/SIGNATURE_CODE.xlsx',
-                'aes/SIGNATURE_CODE.xlsx',
-            ],
         ];
+        if (in_array($mode, ['both', 'collect'])) {
+            $steps = array_merge(
+                $steps,
+                [
+                    'projects' => [
+                        'acs/ALL_PROJECTS.xlsx',
+                        'aes/ALL_PROJECTS.xlsx',
+                    ],
+                    'documents' => [
+                        'acs/DOCUMENTS.xlsx',
+                        'aes/DOCUMENTS.xlsx',
+                    ],
+                    'document_signatures' => [
+                        'acs/DOCUMENTS_Signature.xlsx',
+                        'aes/DOCUMENTS_Signature.xlsx',
+                    ],
+                    'signatures' => [
+                        'acs/SIGNATURE_CODE.xlsx',
+                        'aes/SIGNATURE_CODE.xlsx',
+                    ],
+                ]
+            );
+        }
         foreach ($steps as $method => $files) {
             $name = ucwords(str_replace('_', ' ', $method));
             $stream->send([
@@ -104,13 +122,18 @@ class ManagementController extends Controller
                     'message' => '&nbsp;&nbsp;&nbsp;File #'.($i + 1).' ('.$file.')',
                 ]);
                 $this->$method($stream, $file);
+                // if ($method === 'document_signatures') {
+                //     new RetryDocumentSignatureImport($stream);
+                // }
             }
             $stream->send([
                 'percentage' => 100,
                 'message'    => 'Finished processing files for '.$name,
             ]);
         }
-        // $this->store($stream);
+        if (in_array($mode, ['both', 'store'])) {
+            // $this->store($stream);
+        }
         $stream->stop([
             'message'    => 'Finished importing data',
             'percentage' => 100,
